Decouple base activations from Qwen-Omni specific modules
