+++ models/experimental/tt_symbiote/models/qwen_omni/qwen_omni_modules.py
@@ -1,0 +1,239 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Qwen-Omni specific TTNN modules."""
+
+import os
+
+import torch
+import ttnn
+from models.experimental.tt_symbiote.core.module import TTNNModule
+from models.experimental.tt_symbiote.core.utils import torch_dtype_to_ttnn_dtype, tree_map
+
+
+def _mesh_host_stitch_device_shards(tt_tensor: ttnn.Tensor, mesh_device) -> torch.Tensor | None:
+    """Concat host tensors when each mesh device holds a shard on one logical dimension."""
+    if mesh_device is None or not hasattr(mesh_device, "get_num_devices"):
+        return None
+    nd = int(mesh_device.get_num_devices())
+    if nd <= 1:
+        return None
+    shards = ttnn.get_device_tensors(tt_tensor)
+    if len(shards) != nd:
+        return None
+    local = tuple(int(x) for x in shards[0].shape)
+    for t in (tt_tensor.shape, getattr(tt_tensor, "padded_shape", None)):
+        if t is None:
+            continue
+        logical = tuple(int(x) for x in t)
+        if len(logical) != len(local):
+            continue
+        for d in range(len(logical)):
+            if local[d] != logical[d] and local[d] * nd == logical[d]:
+                parts = [ttnn.to_torch(s).contiguous() for s in shards]
+                return torch.cat(parts, dim=d)
+    return None
+
+
+def _ttnn_mesh_to_torch_one_replica(tt_tensor: ttnn.Tensor, mesh_device) -> torch.Tensor:
+    """Host tensor matching one logical replica (avoid ambiguous mesh compose behavior on replicated tensors)."""
+    if mesh_device is None or mesh_device.get_num_devices() <= 1:
+        return ttnn.to_torch(tt_tensor).contiguous()
+    stitched = _mesh_host_stitch_device_shards(tt_tensor, mesh_device)
+    if stitched is not None:
+        return stitched.contiguous()
+    shards = ttnn.get_device_tensors(tt_tensor)
+    if shards:
+        return ttnn.to_torch(shards[0]).contiguous()
+    composer = ttnn.ConcatMeshToTensor(mesh_device, dim=0)
+    result = ttnn.to_torch(tt_tensor, mesh_composer=composer)
+    lead = int(tt_tensor.shape[0])
+    if result.dim() >= 1 and int(result.shape[0]) > lead:
+        result = result[:lead].contiguous()
+    return result.contiguous()
+
+
+def _upload_bct_replicated(x_t: torch.Tensor, mesh_device):
+    """Upload host tensor with ReplicateTensorToMesh on multi-device meshes."""
+    mesh_mapper = None
+    if mesh_device is not None and hasattr(mesh_device, "get_num_devices") and mesh_device.get_num_devices() > 1:
+        mesh_mapper = ttnn.ReplicateTensorToMesh(mesh_device)
+    return ttnn.from_torch(
+        x_t.contiguous(),
+        dtype=torch_dtype_to_ttnn_dtype(x_t.dtype),
+        layout=ttnn.TILE_LAYOUT,
+        device=mesh_device,
+        mesh_mapper=mesh_mapper,
+    )
+
+
+class TTNNSnakeBeta(TTNNModule):
+    """TTNN SnakeBeta (HF ``SnakeBeta`` in Qwen3-Omni code2wav decoder): x + 1/b * sin^2(x*a)."""
+
+    def __init__(self, in_features: int):
+        super().__init__()
+        self.in_features = in_features
+        self.no_div_by_zero = 0.000000001
+        self.alpha = None
+        self.beta = None
+
+    @classmethod
+    def from_torch(cls, torch_layer, *args, **kwargs):
+        in_features = int(getattr(torch_layer, "in_features", torch_layer.alpha.shape[0]))
+        new_layer = cls(in_features)
+        new_layer._fallback_torch_layer = torch_layer
+        return new_layer
+
+    def move_weights_to_device_impl(self):
+        super().move_weights_to_device_impl()
+        tl = self.torch_layer
+        if tl is None:
+            return
+        w_alpha = tl.alpha.detach().float().contiguous()
+        w_beta = tl.beta.detach().float().contiguous()
+        mesh_mapper = None
+        if self.device is not None and hasattr(self.device, "get_num_devices") and self.device.get_num_devices() > 1:
+            mesh_mapper = ttnn.ReplicateTensorToMesh(self.device)
+        self.alpha = ttnn.from_torch(
+            w_alpha,
+            dtype=ttnn.float32,
+            layout=ttnn.TILE_LAYOUT,
+            device=self.device,
+            mesh_mapper=mesh_mapper,
+            memory_config=ttnn.DRAM_MEMORY_CONFIG,
+        )
+        self.beta = ttnn.from_torch(
+            w_beta,
+            dtype=ttnn.float32,
+            layout=ttnn.TILE_LAYOUT,
+            device=self.device,
+            mesh_mapper=mesh_mapper,
+            memory_config=ttnn.DRAM_MEMORY_CONFIG,
+        )
+
+    def set_output_tensors_config_impl(self, output_tensors):
+        if self.device_state is None or self.device is None or self.device.get_num_devices() <= 1:
+            return super().set_output_tensors_config_impl(output_tensors)
+
+        from models.experimental.tt_symbiote.core.tensor import TorchTTNNTensor
+
+        def _materialize_one_replica(t):
+            if not isinstance(t, TorchTTNNTensor) or t.ttnn_tensor is None:
+                return t
+            tt = t.ttnn_tensor
+            pt = _ttnn_mesh_to_torch_one_replica(tt, self.device)
+            t.elem = pt.contiguous()
+            t.ttnn_tensor = None
+            if getattr(t, "_distributed_tensor_config", None) is not None:
+                t._distributed_tensor_config = None
+            return t
+
+        return tree_map(_materialize_one_replica, output_tensors)
+
+    @staticmethod
+    def _snake_beta_chunk_t() -> int:
+        raw = os.environ.get("TT_SYMBIOTE_SNAKEBETA_CHUNK_T", "4096")
+        try:
+            v = int(raw)
+        except ValueError:
+            v = 4096
+        return max(512, v)
+
+    def _forward_fp32_core(
+        self,
+        input_fp32: ttnn.Tensor,
+        alpha_exp: ttnn.Tensor,
+        reciprocal_beta: ttnn.Tensor,
+    ) -> ttnn.Tensor:
+        x_times_alpha = ttnn.multiply(input_fp32, alpha_exp)
+        sin_result = ttnn.sin(x_times_alpha)
+        sin_squared = ttnn.pow(sin_result, 2.0)
+        scaled_sin = ttnn.multiply(reciprocal_beta, sin_squared)
+        result = ttnn.add(input_fp32, scaled_sin, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+        for t in (x_times_alpha, sin_result, sin_squared, scaled_sin):
+            try:
+                ttnn.deallocate(t)
+            except Exception:
+                pass
+        return result
+
+    def forward(self, input_tensor: ttnn.Tensor) -> ttnn.Tensor:
+        if input_tensor.layout != ttnn.TILE_LAYOUT:
+            input_tensor = ttnn.to_layout(input_tensor, ttnn.TILE_LAYOUT, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+
+        out_dtype = input_tensor.dtype
+        shape = tuple(int(s) for s in input_tensor.shape)
+        if len(shape) != 3:
+            raise ValueError(f"TTNNSnakeBeta expects rank-3 [B,C,T], got shape {shape}")
+        b, c, t_len = shape
+
+        chunk_t = self._snake_beta_chunk_t()
+        alpha_expanded = ttnn.unsqueeze(self.alpha, 0)
+        alpha_expanded = ttnn.unsqueeze(alpha_expanded, -1)
+        beta_expanded = ttnn.unsqueeze(self.beta, 0)
+        beta_expanded = ttnn.unsqueeze(beta_expanded, -1)
+        alpha_exp = ttnn.exp(alpha_expanded)
+        try:
+            ttnn.deallocate(alpha_expanded)
+        except Exception:
+            pass
+        beta_exp = ttnn.exp(beta_expanded)
+        beta_plus_eps = ttnn.add(beta_exp, self.no_div_by_zero)
+        reciprocal_beta = ttnn.reciprocal(beta_plus_eps)
+        try:
+            ttnn.deallocate(beta_expanded)
+            ttnn.deallocate(beta_exp)
+            ttnn.deallocate(beta_plus_eps)
+        except Exception:
+            pass
+
+        if t_len <= chunk_t:
+            if out_dtype != ttnn.float32:
+                input_fp32 = ttnn.typecast(input_tensor, ttnn.float32)
+            else:
+                input_fp32 = input_tensor
+            result = self._forward_fp32_core(input_fp32, alpha_exp, reciprocal_beta)
+            if out_dtype != ttnn.float32:
+                if input_fp32 is not input_tensor:
+                    try:
+                        ttnn.deallocate(input_fp32)
+                    except Exception:
+                        pass
+                result = ttnn.typecast(result, out_dtype)
+            try:
+                ttnn.deallocate(alpha_exp)
+                ttnn.deallocate(reciprocal_beta)
+            except Exception:
+                pass
+            return result
+
+        out_chunks = []
+        for t0 in range(0, t_len, chunk_t):
+            t1 = min(t0 + chunk_t, t_len)
+            sl = ttnn.slice(input_tensor, (0, 0, t0), (b, c, t1))
+            if out_dtype != ttnn.float32:
+                sl_fp32 = ttnn.typecast(sl, ttnn.float32)
+            else:
+                sl_fp32 = sl
+            res_fp32 = self._forward_fp32_core(sl_fp32, alpha_exp, reciprocal_beta)
+            if out_dtype != ttnn.float32:
+                out_chunks.append(ttnn.typecast(res_fp32, out_dtype))
+            else:
+                out_chunks.append(res_fp32)
+
+        torch_parts = []
+        mesh_dev = self.device
+        for ch in out_chunks:
+            torch_parts.append(_ttnn_mesh_to_torch_one_replica(ch, mesh_dev))
+            try:
+                ttnn.deallocate(ch)
+            except Exception:
+                pass
+        merged_torch = torch.cat(torch_parts, dim=2)
+        try:
+            ttnn.deallocate(alpha_exp)
+            ttnn.deallocate(reciprocal_beta)
+        except Exception:
+            pass
+        return _upload_bct_replicated(merged_torch, mesh_dev)
